added new file, hibp.go
