+++ stringutils.go
@@ -8,11 +8,8 @@
 package crunchy
 
 import (
-	"crypto/sha1"
 	"encoding/hex"
 	"hash"
-	"io/ioutil"
-	"net/http"
 	"strings"
 	"unicode"
 	"unicode/utf8"
@@ -72,34 +69,4 @@ func hashsum(s string, hasher hash.Hash) string {
 	hasher.Reset()
 	hasher.Write([]byte(s))
 	return hex.EncodeToString(hasher.Sum(nil))
-}
-
-func foundInHIBP(s string) error {
-	h := sha1.New()
-	h.Write([]byte(s))
-	result := hex.EncodeToString(h.Sum(nil))
-
-	firstFive := result[0:5]
-	restOfHash := strings.ToUpper(result[5:])
-
-	url := "https://api.pwnedpasswords.com/range/" + firstFive
-
-	resp, err := http.Get(url)
-	if err != nil {
-		return err
-	}
-
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-
-	if strings.Index(string(body), restOfHash) > -1 {
-		return ErrFoundHIBP
-	}
-
-	return nil
-
 }
